ASoC: codecs: lpass-wsa-macro: use ports as the DAIs for playback
AIF1_PB and AIF_MIX1_PB don't correspond to any part of the HW, they are
"virtual" inputs that are configured with mixer settings. This
configuration shouldn't be part of this driver, and AFAIK its not used in
any useful way upstream so it can just be removed.

The inputs to the WSA codec are the ports (RX0 to RX8), so remove AIF1_PB
and AIF_MIX1_PB and expose the ports as the DAIs.

Note the DAIs correspond to an INTF of the DMA engine. For example, RX0/RX1
come from the same INTF so they are part of the same DAI. (they must have
the same rate).

The common use-case is to use AIF1_PB with RX0/RX1, and AIF_MIX1_PB with
RX_MIX0/RX_MIX1, so this change should not break things for current users
(except 'WSA RX0 MUX', etc. settings will longer exist).

Signed-off-by: Jonathan Marek <jonathan@marek.ca>